requirements.txt
discord.py == 2.3.2
yt-dlp